+++ packages/rma-frontend/src/app/sales-ui/sales/sales.page.ts
@@ -6,7 +6,6 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { SalesInvoiceDataSource } from './sales-invoice-datasource';
 import { Router, NavigationEnd, ActivatedRoute } from '@angular/router';
-import { VIEW_SALES_INVOICE_PAGE_URL } from '../../constants/url-strings';
 import { map, filter, startWith, switchMap } from 'rxjs/operators';
 import { FormControl, FormGroup } from '@angular/forms';
 import {
@@ -48,12 +47,14 @@ export class SalesPage implements OnInit {
     'posting_time',
     'customer_name',
     'total',
+    'delivered_percent',
     'due_amount',
     'remarks',
     'territory',
     'created_by',
     'delivered_by',
   ];
+  bufferValue = 70;
   invoiceStatus: string[] = [
     'Draft',
     'Completed',
@@ -348,12 +349,6 @@ export class SalesPage implements OnInit {
         : this.sortQuery;
 
     this.dataSource.loadItems(this.sortQuery, undefined, undefined, query);
-  }
-
-  navigateBasedOnRoles(row) {
-    this.router.navigateByUrl(
-      `/sales/${VIEW_SALES_INVOICE_PAGE_URL}/${row.uuid}`,
-    );
   }
 
   navigateBack() {