+++ packages/rma-frontend/src/app/stock-entry/stock-availability/stock-availability.page.ts
@@ -5,11 +5,11 @@ import { FormControl, FormGroup } from '@angular/forms';
 
 import { StockAvailabilityDataSource } from './stock-availability-datasource';
 import { SalesService } from '../../sales-ui/services/sales.service';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { ValidateInputSelected } from '../../common/pipes/validators';
 import { startWith, switchMap } from 'rxjs/operators';
 import { ActivatedRoute } from '@angular/router';
-import { StockEntryService } from '../services/stock-entry/stock-entry.service';
+import { WAREHOUSES } from '../../constants/app-string';
 
 @Component({
   selector: 'app-stock-availability',
@@ -21,13 +21,22 @@ export class StockAvailabilityPage implements OnInit {
 
   dataSource: StockAvailabilityDataSource;
   defaultCompany: string;
-  displayedColumns = ['item', 'warehouse', 'actual_qty'];
+  displayedColumns = [
+    'excel_item_name',
+    'item_code',
+    'excel_item_group',
+    'excel_item_brand',
+    'warehouse',
+    'actual_qty',
+  ];
   filters: any = [];
   countFilter: any = {};
   stockAvailabilityForm: FormGroup;
   filteredStockAvailabilityList: Observable<any>;
   filteredWarehouseList: Observable<any[]>;
   validateInput: any = ValidateInputSelected;
+  filteredItemGroupList: Observable<any>;
+  filteredItemBrandList: Observable<any>;
 
   get f() {
     return this.stockAvailabilityForm.controls;
@@ -36,7 +45,6 @@ export class StockAvailabilityPage implements OnInit {
   constructor(
     private readonly location: Location,
     private readonly salesService: SalesService,
-    private readonly stockEntryService: StockEntryService,
     private route: ActivatedRoute,
   ) {}
 
@@ -61,22 +69,57 @@ export class StockAvailabilityPage implements OnInit {
       .get('warehouse')
       .valueChanges.pipe(
         startWith(''),
-        switchMap(() => {
-          return this.stockEntryService.getWarehouseList();
+        switchMap(value => {
+          return this.salesService.getStore().getItemAsync(WAREHOUSES, value);
         }),
       );
+
+    this.filteredItemGroupList = this.stockAvailabilityForm
+      .get('excel_item_group')
+      .valueChanges.pipe(
+        startWith(''),
+        switchMap(value => {
+          return this.salesService.getItemGroupList(value);
+        }),
+        switchMap(data => {
+          return of(data);
+        }),
+      );
+
+    this.filteredItemBrandList = this.stockAvailabilityForm
+      .get('excel_item_brand')
+      .valueChanges.pipe(
+        startWith(''),
+        switchMap(value => {
+          return this.salesService.getItemBrandList(value);
+        }),
+        switchMap(data => {
+          return of(data);
+        }),
+      );
+  }
+
+  getItemBrandOption(option) {
+    if (option) {
+      if (option.brand) {
+        return `${option.brand}`;
+      }
+      return option.brand;
+    }
   }
 
   createFormGroup() {
     this.stockAvailabilityForm = new FormGroup({
       itemName: new FormControl(),
       warehouse: new FormControl(),
+      excel_item_group: new FormControl(),
+      excel_item_brand: new FormControl(),
     });
   }
 
   getStockAvailabilityOption(option) {
     if (option) {
-      return option.name;
+      return option.item_name;
     }
   }
 
@@ -89,6 +132,8 @@ export class StockAvailabilityPage implements OnInit {
   clearFilters() {
     this.f.itemName.setValue('');
     this.f.warehouse.setValue('');
+    this.f.excel_item_brand.setValue('');
+    this.f.excel_item_group.setValue('');
     this.setFilter();
   }
 
@@ -118,17 +163,46 @@ export class StockAvailabilityPage implements OnInit {
     }
 
     if (this.f.warehouse.value) {
+      this.filters.push(['warehouse', 'like', `%${this.f.warehouse.value}%`]);
+      this.countFilter.warehouse = ['like', `%${this.f.warehouse.value}%`];
+    }
+
+    if (this.f.excel_item_group.value) {
       this.filters.push([
-        'warehouse',
+        'excel_item_group',
         'like',
-        `%${this.f.warehouse.value.name}%`,
+        `%${this.f.excel_item_group.value.name}%`,
       ]);
-      this.countFilter.warehouse = ['like', `%${this.f.warehouse.value.name}%`];
+      this.countFilter.excel_item_group = [
+        'like',
+        `%${this.f.excel_item_group.value.name}%`,
+      ];
+    }
+
+    if (this.f.excel_item_brand.value) {
+      this.filters.push([
+        'excel_item_brand',
+        'like',
+        `%${this.f.excel_item_brand.value.brand}%`,
+      ]);
+      this.countFilter.excel_item_brand = [
+        'like',
+        `%${this.f.excel_item_brand.value.brand}%`,
+      ];
     }
     this.dataSource.loadItems(0, 30, this.filters, this.countFilter);
   }
 
   navigateBack() {
     this.location.back();
+  }
+
+  getItemGroupOption(option) {
+    if (option) {
+      if (option.item_group_name) {
+        return `${option.item_group_name}`;
+      }
+      return option.item_group_name;
+    }
   }
 }