+++ packages/rma-server/src/constants/agenda-job.ts
@@ -1,3 +1,5 @@
+import { SYSTEM_MANAGER } from './app-strings';
+
 export const AGENDA_JOB_METADATA = {
   name: '',
   type: 'single',
@@ -14,7 +16,10 @@ export const AGENDA_JOB_METADATA = {
   lastFinishedAt: new Date(),
 };
 
-export function getParsedPostingDate(payload) {
+export function getParsedPostingDate(payload: {
+  posting_date?: string;
+  posting_time?: string;
+}) {
   let date: Date;
   try {
     date = new Date(
@@ -34,4 +39,27 @@ export function reversePosingDate(date: string) {
   return splitDate[0].length === 4
     ? splitDate.join()
     : splitDate.reverse().join('-');
+}
+
+export function getUserPermissions(req: {
+  token: PermissionStateInterface;
+}): PermissionStateInterface {
+  const userPermissions: PermissionStateInterface = {
+    name: req.token.name,
+    fullName: req.token.fullName,
+    email: req.token.email,
+  };
+  if (req.token.roles.includes(SYSTEM_MANAGER)) {
+    return userPermissions;
+  }
+  return { ...userPermissions, ...req.token };
+}
+
+export interface PermissionStateInterface {
+  name: string;
+  fullName: string;
+  email: string;
+  warehouses?: string[];
+  territories?: string[];
+  roles?: string[];
 }