+++ packages/rma-warranty/src/app/common/app-common.module.ts
@@ -2,11 +2,12 @@ import { NgModule } from '@angular/core';
 import { CurrencyFormatPipe } from './pipes/currency-format.pipe';
 import { CommonModule, CurrencyPipe } from '@angular/common';
 import { SerialsService } from './helpers/serials/serials.service';
+import { CountDownPipe } from './pipes/countdown.pipe';
 
 @NgModule({
   imports: [CommonModule],
-  declarations: [CurrencyFormatPipe],
+  declarations: [CurrencyFormatPipe, CountDownPipe],
   providers: [CurrencyPipe, SerialsService],
-  exports: [CurrencyFormatPipe],
+  exports: [CurrencyFormatPipe, CountDownPipe],
 })
 export class AppCommonModule {}